Fix circular dependencies
Note: Typescript files with only types, are used only in compilation
phase and not going to be imported on compiled javascript.

Closes #1074
