+++ src/index.ts
@@ -65,7 +65,7 @@ import ApolloClient from './ApolloClient';
 
 import {
   ApolloQueryResult,
-} from './core/QueryManager';
+} from './core/types';
 
 import {
   toIdValue,
