+++ src/optimistic-data/store.ts
@@ -14,10 +14,10 @@ import {
 } from '../data/storeUtils';
 
 import {
-  getDataWithOptimisticResults,
   Store,
 } from '../store';
 
+import assign from 'lodash/assign';
 import pick from 'lodash/pick';
 
 // a stack of patches of new or changed documents
@@ -27,6 +27,14 @@ export type OptimisticStore = {
 }[];
 
 const optimisticDefaultState: any[] = [];
+
+export function getDataWithOptimisticResults(store: Store): NormalizedCache {
+  if (store.optimistic.length === 0) {
+    return store.data;
+  }
+  const patches = store.optimistic.map(opt => opt.data);
+  return assign({}, store.data, ...patches) as NormalizedCache;
+}
 
 export function optimistic(
   previousState = optimisticDefaultState,
@@ -49,7 +57,7 @@ export function optimistic(
     const fakeStore = {
       ...store,
       optimistic: previousState,
-    } as Store;
+    };
     const optimisticData = getDataWithOptimisticResults(fakeStore);
     const fakeDataResultState = data(
       optimisticData,
